chore: Create albums summary function to use on "Albums" screen
