+++ mobile/src/data/album/queries.ts
@@ -1,0 +1,9 @@
+import { useQuery } from "@tanstack/react-query";
+
+import { queries as q } from "~/queries/keyStore";
+
+//#region Queries
+export function useAlbums() {
+  return useQuery({ ...q.albums.all });
+}
+//#endregion
